rename: StudyOnceQuestion 엔티티명, 변수명, 메서드명 변경
