+++ src/main/java/com/example/demo/repository/StudyOnceCommentRepository.java
@@ -1,0 +1,8 @@
+package com.example.demo.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.demo.domain.StudyOnceComment;
+
+public interface StudyOnceQuestionRepository extends JpaRepository<StudyOnceComment, Long> {
+}
