--- src/main/java/com/example/demo/repository/StudyOnceQuestionRepository.java
@@ -1,8 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.demo.domain.StudyOnceQuestion;
-
-public interface StudyOnceQuestionRepository extends JpaRepository<StudyOnceQuestion, Long> {
-}
