+++ src/main/java/com/example/demo/service/StudyOnceCommentServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.domain.MemberImpl;
+import com.example.demo.domain.StudyOnceComment;
 import com.example.demo.domain.StudyOnceImpl;
-import com.example.demo.domain.StudyOnceQuestion;
 import com.example.demo.dto.StudyOnceQuestionRequest;
 import com.example.demo.dto.StudyOnceQuestionUpdateRequest;
 import com.example.demo.exception.CafegoryException;
@@ -28,18 +28,18 @@ public class StudyOnceQuestionServiceImpl implements StudyOnceQuestionService {
 
 	@Override
 	public Long saveQuestion(Long memberId, Long studyOnceId, StudyOnceQuestionRequest request) {
-		StudyOnceQuestion question = StudyOnceQuestion.builder()
+		StudyOnceComment question = StudyOnceComment.builder()
 			.content(request.getContent())
 			.member(findMemberById(memberId))
 			.studyOnce(findStudyOnceById(studyOnceId))
 			.build();
-		StudyOnceQuestion savedQuestion = studyOnceQuestionRepository.save(question);
+		StudyOnceComment savedQuestion = studyOnceQuestionRepository.save(question);
 		return savedQuestion.getId();
 	}
 
 	@Override
 	public void updateQuestion(Long memberId, Long studyOnceQuestionId, StudyOnceQuestionUpdateRequest request) {
-		StudyOnceQuestion question = findStudyOnceQuestionById(studyOnceQuestionId);
+		StudyOnceComment question = findStudyOnceQuestionById(studyOnceQuestionId);
 		if (!isPersonWhoAskedQuestion(memberId, studyOnceQuestionId)) {
 			throw new CafegoryException(STUDY_ONCE_QUESTION_PERMISSION_DENIED);
 		}
@@ -48,30 +48,30 @@ public class StudyOnceQuestionServiceImpl implements StudyOnceQuestionService {
 
 	@Override
 	public boolean isPersonWhoAskedQuestion(Long memberId, Long studyOnceQuestionId) {
-		StudyOnceQuestion question = findStudyOnceQuestionById(studyOnceQuestionId);
+		StudyOnceComment question = findStudyOnceQuestionById(studyOnceQuestionId);
 		return question.isPersonAsked(findMemberById(memberId));
 	}
 
 	@Override
 	public Long saveReply(Long memberId, Long studyOnceId, Long parentStudyOnceQuestionId,
 		StudyOnceQuestionRequest request) {
-		StudyOnceQuestion reply = StudyOnceQuestion.builder()
+		StudyOnceComment reply = StudyOnceComment.builder()
 			.content(request.getContent())
 			.member(findMemberById(memberId))
 			.studyOnce(findStudyOnceById(studyOnceId))
 			.parent(findStudyOnceQuestionById(parentStudyOnceQuestionId))
 			.build();
-		StudyOnceQuestion savedReply = studyOnceQuestionRepository.save(reply);
+		StudyOnceComment savedReply = studyOnceQuestionRepository.save(reply);
 		return savedReply.getId();
 	}
 
 	@Override
 	public void deleteQuestion(Long studyOnceQuestionId) {
-		StudyOnceQuestion question = findStudyOnceQuestionById(studyOnceQuestionId);
+		StudyOnceComment question = findStudyOnceQuestionById(studyOnceQuestionId);
 		studyOnceQuestionRepository.delete(question);
 	}
 
-	private StudyOnceQuestion findStudyOnceQuestionById(Long studyOnceQuestionId) {
+	private StudyOnceComment findStudyOnceQuestionById(Long studyOnceQuestionId) {
 		return studyOnceQuestionRepository.findById(studyOnceQuestionId)
 			.orElseThrow(() -> new CafegoryException(STUDY_ONCE_QUESTION_NOT_FOUND));
 	}
